+++ Achievements/Childrens-Week/Pets.lua
@@ -2,9 +2,10 @@ RXPGuides.RegisterGuide([[
 #wotlk
 << Alliance
 #name Pets Childrens Week
-#displayname Children's Week
+#displayname Pets
 #version 1
-#group Pets
+#group Achievements
+#subgroup Children's Week
 
 --|cFFFCDC00Note|r
 --|cFF00FF25Friendly|r
@@ -29,6 +30,7 @@ step
     .accept 1479
     .use 18598
 step
+    .itemcount 18598,1
     >>Buy icecream from |cFF00FF25Emmithue Smails|r in Stormwind.
     .collect 7228,1
     .goto Stormwind City,61.72,75.80
@@ -124,7 +126,6 @@ step
     .use 31881
 step
     >>Take the |cFF00FF25Dornaa|r to the Meeting Stone in the middle of Auchindoun.
-    >>Click the portal to help |cFF00FF25Dornaa|r summon an adventurer. 
     .complete 10950,1
     .goto Terokkar Forest,39.71,64.59
     .use 31881
@@ -165,6 +166,11 @@ step
     .goto Nagrand,31.47,57.59
     .target Jheel
 step
+    >>Talk to |cFF00FF25Dornaa|r.
+    .accept 10962
+    .accept 10956
+    .use 31881
+step
     #completewith next
     .zone Dalaran >>Teleport to Dalaran << Mage
     .zone Dalaran >>Head to Dalaran << !Mage
@@ -178,10 +184,8 @@ step
     .goto Tanaris,63.6,57.6
     .target Alurmi
 step
-    >>Approach |cFF00FF25Zaladormu|r and use the |cFF00BCD4Draenei Orphan Whistle|r to summon |cFF00FF25Dornaa|r.
-    >>An RP will play out, then give |cFF00FF25Dornaa|r the |cFF00BCD4Toy Dragon|r.
+    >>Approach |cFF00FF25Zaladormu|r and use the |cFF00BCD4Draenei Orphan Whistle|r to summon |cFF00FF25Dornaa|r. Then give her the |cFF00BCD4Toy Dragon|r.
     .turnin 10962
-    .accept 10956
     .goto Tanaris,60.36,57.42
     .use 31881
 step
@@ -206,11 +210,12 @@ step
     .use 31881
     .target Farseer Nobundo
 step
-    >>Talk to |cFF00FF25Farseer Nobundo|r.
+    >>Talk to |cFF00FF25Farseer Nobundo|r then |cFF00FF25Dornaa|r.
     .turnin 10968
     .accept 10966
     .goto The Exodar,29.90,32.85
     .target Farseer Nobundo
+    .target Draenei Orphan
 step
     #completewith next
     .zone Shattrath City >> Teleport to Shattrath << Mage
@@ -239,49 +244,250 @@ step
 -- Curious Wolvar Pup
 -- ******************
 step
-    >>.isOnQuest 13927
+    .isOnQuest 13927
     >>|cFFFCDC00Curious Wolvar Pup|r
+    .complete 13927,1
+    .goto Dalaran,49.4,63.2
+    .target Orphan Matron Aria
+step
+    .itemcount 46396,1
     >>Use the |cFF00BCD4Wolvar Orphan Whistle|r.
     >>Talk to the |cFF00FF25Wolvar Orphan (Kekek)|r.
     .turnin 13927
     .accept 13930
-    .accept 13934
     .accept 13951
+    .accept 13934
     .use 46396
 step
-    >>.isOnQuest 13930
+    .itemcount 46396,1
     #completewith next
     .fly Westfall Brigade
     .goto Dalaran,71.93,45.75
 step
-    >>.isOnQuest 13930
+    .itemcount 46396,1
     >>Take |cFF00FF25Kekek|r to visit the giant tree in Grizzlemaw.
     .complete 13930,1
-    .goto Grizzly Hills,50.8,42.8
+    .goto Grizzly Hills,50.83,42.66
+    .use 46396
+step
+    .itemcount 46396,1
+    >>Talk to |cFF00FF25Kekek|r.
+    .turnin 13930
+    .use 46396
+step
+    .itemcount 46396,1
+    >>Take |cFF00FF25Kekek|r to visit Snowfall Glade in Dragonblight.
+    .complete 13951,1
+    .goto Dragonblight,49.74,63.29
+    .use 46396
+step
+    .itemcount 46396,1
+    >>Talk to |cFF00FF25Kekek|r.
+    .turnin 13951
+    .use 46396
+step
+    .itemcount 46396,1
+    >>Take |cFF00FF25Kekek|r to visit Bronze Dragonshrine.
+    .complete 13934,1
+    .goto Dragonblight,71.1,41
+    .use 46396
+step
+    .itemcount 46396,1
+    >>Talk to |cFF00FF25Kekek|r.
+    .turnin 13934
+    .accept 13955
+    .accept 13957
+    .use 46396
+step
+    .itemcount 46396,1
+    >>Take |cFF00FF25Kekek|r to visit Alexstrasza at the top of Wyrmrest Temple.
+    .complete 13955,1
+    .goto Dragonblight,59.82,54.61
+    .use 46396
+step
+    .itemcount 46396,1
+    >>Talk to |cFF00FF25Kekek|r.
+    .turnin 13955
+    .use 46396
+step
+    .itemcount 46396,1
+    #completewith next
+    .fly River's Heart >> Fly or Wormhole to Sholazar Basin
+    .goto Dragonblight,60.33,51.50
+step
+    .itemcount 46396,1
+    >>Take |cFF00FF25Kekek|r to visit Hemet Nesingwary at the Nesingwary Base Camp.
+    .complete 13957,1
+    .goto Sholazar Basin,27.1,58.72
+    .use 46396
+step
+    .itemcount 46396,1
+    >>Talk to |cFF00FF25Kekek|r.
+    .turnin 13957
+    .accept 13938
+    .use 46396
+step
+    .itemcount 46396,1
+    #completewith next
+    .zone Dalaran >>Teleport to Dalaran << Mage
+    .zone Dalaran >>Head to Dalaran << !Mage
+step
+    .itemcount 46396,1
+    #completewith next
+    >>Head to The Wonderworks toy shop.
+    >>Talk to |cFF00FF25Jepetto Joybuzz|r and buy:
+    .collect 46693,1
+    .goto Dalaran,44.8,46.2
+    .target Jepetto Joybuzz
+step
+    .itemcount 46396,1
+    >>Throw the |cFF00BCD4Small Paper Zeppelin|r to |cFF00FF25Kekek|r.
+    .complete 13938,1
+    .use 46396
+    .use 46693
+step
+    .itemcount 46396,1
+    >>Talk to |cFF00FF25Kekek|r.
+    .turnin 13938
+    .accept 13960
+    .use 46396
+step
+    .itemcount 46396,1
+    >>Talk to |cFF00FF25Orphan Matron Aria|r in Dalaran.
+    .turnin 13960
+    .goto Dalaran,49.4,63.2
+    .target Orphan Matron Aria
 
 -- *************************
 -- Curious Oracle Hatchling
 -- *************************
 step
-    >>.isOnQuest 13926
     >>|cFFFCDC00Curious Oracle Hatchling|r
+    .isOnQuest 13926
+    .complete 13926,1
+    .goto Dalaran,49.4,63.2
+    .target Orphan Matron Aria
+step
+    .itemcount 46397,1
     >>Use the |cFF00BCD4Oracle Orphan Whistle|r.
     >>Talk to the |cFF00FF25Oracle Orphan (Roo)|r.
     .turnin 13926
     .accept 13929
-    .accept 13933
     .accept 13950
+    .accept 13933
     .use 46397
 step
-    >>.isOnQuest 13929
+    .itemcount 46397,1
     #completewith next
     .fly Westfall Brigade
     .goto Dalaran,71.93,45.75
 step
-    >>.isOnQuest 13929
+    .itemcount 46397,1
     >>Take |cFF00FF25Roo|r to visit the giant tree in Grizzlemaw.
     .complete 13929,1
-    .goto Grizzly Hills,50.8,42.8
+    .goto Grizzly Hills,50.83,42.66
+    .use 46397
+step
+    .itemcount 46397,1
+    >>Talk to |cFF00FF25Roo|r.
+    .turnin 13929
+    .use 46397
+step
+    .itemcount 46397,1
+    #completewith next
+    .zone Borean Tundra >>Head to Borean Tundra
+step
+    .itemcount 46397,1
+    >>Take |cFF00FF25Roo|r to visit Winterfin Retreat.
+    .complete 13950,1
+    .goto Borean Tundra,43.5,13.6
+    .use 46397
+step
+    .itemcount 46397,1
+    >>Talk to |cFF00FF25Roo|r.
+    .turnin 13950
+    .use 46397
+step
+    .itemcount 46397,1
+    #completewith next
+    .zone Dragonblight >>Head to Dragonblight
+step
+    .itemcount 46397,1
+    >>Take |cFF00FF25Roo|r to visit the Bronze Dragonshrine.
+    .complete 13933,1
+    .goto Dragonblight,72.74,36.58
+    .use 46397
+step
+    .itemcount 46397,1
+    >>Talk to |cFF00FF25Roo|r.
+    .turnin 13933
+    .accept 13954
+    .accept 13956
+    .use 46397
+step
+    .itemcount 46397,1
+    >>Take |cFF00FF25Roo|r to visit Alexstrasza at the top of Wyrmrest Temple.
+    .complete 13954,1
+    .goto Dragonblight,59.82,54.61
+    .use 46397
+step
+    .itemcount 46397,1
+    >>Talk to |cFF00FF25Roo|r.
+    .turnin 13954
+    .use 46397
+step
+    .itemcount 46397,1
+    #completewith next
+    .fly River's Heart >> Fly or Wormhole to Sholazar Basin
+    .goto Dragonblight,60.33,51.50
+step
+    .itemcount 46397,1
+    #completewith next
+    .goto Sholazar Basin,40.24,82.69,5 >>Travel through the waygate in Scholazar Basin to Un'Goro Crater.
+step
+    .itemcount 46397,1
+    >>Take |cFF00FF25Roo|r to visit the Etymidian in Un'Goro.
+    .complete 13956,1
+    .goto Un'Goro Crater,47.51,9.24
+    .use 46397
+step
+    .itemcount 46397,1
+    >>Talk to |cFF00FF25Roo|r.
+    .turnin 13956
+    .accept 13937
+    .use 46397
+step
+    .itemcount 46397,1
+    #completewith next
+    .zone Dalaran >>Teleport to Dalaran << Mage
+    .zone Dalaran >>Head to Dalaran << !Mage
+step
+    .itemcount 46397,1
+    #completewith next
+    >>Head to The Wonderworks toy shop.
+    >>Talk to |cFF00FF25Jepetto Joybuzz|r and buy:
+    .collect 46693,1
+    .goto Dalaran,44.8,46.2
+    .target Jepetto Joybuzz
+step
+    .itemcount 46397,1
+    >>Throw the |cFF00BCD4Small Paper Zeppelin|r to |cFF00FF25Roo|r.
+    .complete 13937,1
+    .use 46397
+    .use 46693
+step
+    .itemcount 46397,1
+    >>Talk to |cFF00FF25Roo|r.
+    .turnin 13937
+    .accept 13959
+    .use 46397
+step
+    .itemcount 46397,1
+    >>Talk to |cFF00FF25Orphan Matron Aria|r in Dalaran.
+    .turnin 13959
+    .goto Dalaran,49.4,63.2
+    .target Orphan Matron Aria
 step
     >>CONGRATULATIONS!!! You've collected all possible pets for this year.
+    >>|cFFFCDC00Note|r: Your Northrend pet will arrive in the mail.
 ]])