+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## Versions
 
-## 1.6.0
-- Add Pets section
+## 2.0.0
+- Re-arrange files and menus
 - Add Guide for Children's Week pets
+- Small updates to "For The Children" guide
 - README updates
 
 ## 1.5.0